Refactor inbox tests to use new worker name functions
